+++ CHANGELOG.md
@@ -16,6 +16,10 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+### Fixed
+
+- Fixed error when returning responses from deleted users (which contains user_id=None). ([#57](https://github.com/argilla-io/argilla-server/pull/57))
+
 ## [1.25.0](https://github.com/argilla-io/argilla-server/compare/v1.24.0...v1.25.0)
 
 >[!IMPORTANT]